// types.ts
export interface Service {
  id: number;
  name: string;
  imageUrl: string;
}

export interface Testimonial {
  id: number;
  quote: string;
  author: string;
  location: string;
}

export interface NewsLogo {
  id: number;
  name: string;
  logoUrl: string;
}

export interface InstagramPost {
    id: number;
    imageUrl: string;
}